+++ braindecode/datautil/__init__.py
@@ -47,3 +47,6 @@ def __getattr__(name):
         return windowers.__dict__[name]
 
     raise AttributeError("No possible import named " + name)
+
+
+__all__ = ["load_concat_dataset", "save_concat_dataset", "_check_save_dir_empty"]