+++ braindecode/models/util.py
@@ -5,11 +5,7 @@
 import inspect
 from pathlib import Path
 
-import numpy as np
 import pandas as pd
-import torch
-from scipy.special import log_softmax
-from sklearn.utils import deprecated
 
 import braindecode.models as models